Revert "Revert "Revert "monet: colors: Switch to single-precision floating-point math"""
This reverts commit 6aabce01062886ec8e017870de9707224e79975e.

+++ app/src/main/java/dev/kdrag0n/android12ext/core/xposed/hooks/SystemUIHooks.kt
@@ -59,7 +59,7 @@ class SystemUIHooks(
 
     fun applyThemeOverlayController(
         isGoogle: Boolean,
-        chromaMultiplier: Float,
+        chromaMultiplier: Double,
         accurateShades: Boolean,
         colorOverride: Int?,
     ) {
